Basic nests are working, but it ends up putting them too close to your home

+++ core/src/org/noses/games/homedefense/client/Way.java
@@ -112,7 +112,7 @@ public class Way {
         return endNode;
     }
 
-    public float getDistance() {
+    public double getDistance() {
         return firstNode().distanceFrom(lastNode());
     }
